[#5]wrap up project
