+++ MessagerPractice/Controllers/ChatViewController.swift
@@ -14,61 +14,8 @@ import SDWebImage
 import AVFoundation
 import AVKit
 
-struct Message : MessageType {
-    public var sender: SenderType
-    public var messageId: String
-    public var sentDate: Date
-    public var kind: MessageKind
-}
 
-extension MessageKind {
-    var messageKindString : String {
-        switch self{
-            
-        case .text(_):
-            return "text"
-        case .attributedText(_):
-            return "attributed_text"
-        case .photo(_):
-            return "photo"
-        case .video(_):
-            return "video"
-        case .location(_):
-            return "location"
-        case .emoji(_):
-            return "emoji"
-        case .audio(_):
-            return "audio"
-        case .contact(_):
-            return "contact"
-        case .linkPreview(_):
-            return "linkPreview"
-        case .custom(_):
-            return "custom"
-        }
-    }
-}
-
-struct Sender : SenderType {
-    public var photoURL: String
-    public var senderId: String
-    public var displayName: String
-
-}
-
-struct Media : MediaItem {
-    var url: URL?
-    var image: UIImage?
-    var placeholderImage: UIImage
-    var size: CGSize
-}
-
-struct Location : LocationItem {
-    var location: CLLocation
-    var size: CGSize
-}
-
-class ChatViewController: MessagesViewController {
+final class ChatViewController: MessagesViewController {
 
     private var senderPhotoURL : URL?
     private var otherUserPhotoURL : URL?
@@ -99,8 +46,8 @@ class ChatViewController: MessagesViewController {
     
     
     init(with email : String,id:String?) {
-        self.otherUserEmail = email
-        self.conversationId = id
+        otherUserEmail = email
+        conversationId = id
         super.init(nibName: nil, bundle: nil)
         
     }
@@ -272,8 +219,8 @@ extension ChatViewController : UIImagePickerControllerDelegate,UINavigationContr
         guard
             let messageId = createMessageId(),
             let conversationID = conversationId,
-            let name = self.title,
-            let selfSender = self.selfSender else {
+            let name = title,
+            let selfSender = selfSender else {
                 
                 return
             }
@@ -522,7 +469,7 @@ extension ChatViewController : MessageCellDelegate {
             guard let imageUrl = media.url else {return}
             let vc = PhotoViewerViewController(with: imageUrl)
             
-            self.navigationController?.pushViewController(vc, animated: true)
+            navigationController?.pushViewController(vc, animated: true)
         case .video(let media):
             guard let videoUrl = media.url else {return}
