+++ MessagerPractice/Controllers/Login/LoginViewController.swift
@@ -12,7 +12,7 @@ import GoogleSignIn
 import Firebase
 import JGProgressHUD
 
-class LoginViewController: UIViewController {
+final class LoginViewController: UIViewController {
     
     private let spinner = JGProgressHUD(style: .dark)
