+++ MessagerPractice/Controllers/Login/RegisterViewController.swift
@@ -9,7 +9,7 @@ import UIKit
 import FirebaseAuth
 import JGProgressHUD
 
-class RegisterViewController: UIViewController {
+final class RegisterViewController: UIViewController {
     
     private let spinner = JGProgressHUD(style: .dark)
     
@@ -293,7 +293,7 @@ extension RegisterViewController : UIImagePickerControllerDelegate, UINavigation
         
         guard let selectedImage = info[UIImagePickerController.InfoKey.editedImage] as? UIImage else {return}
         
-        self.imageView.image = selectedImage
+        imageView.image = selectedImage
     }
     func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
         picker.dismiss(animated: true, completion: nil)
