+++ MessagerPractice/Controllers/NewConversationViewController.swift
@@ -8,7 +8,7 @@ import UIKit
 import JGProgressHUD
 import SwiftUI
 
-class NewConversationViewController: UIViewController {
+final class NewConversationViewController: UIViewController {
 
     public var completion: ((SearchResult) -> Void)?
     
@@ -105,7 +105,7 @@ extension NewConversationViewController : UISearchBarDelegate {
         searchBar.resignFirstResponder()
         results.removeAll()
         spinner.show(in: view)
-        self.searchUsers(query: text)
+        searchUsers(query: text)
     }
     
     func searchUsers(query : String) {
@@ -139,9 +139,9 @@ extension NewConversationViewController : UISearchBarDelegate {
             return
         }
         let safeEmail = DatabaseManager.safeEmail(email: currentUserEmail)
-        self.spinner.dismiss()
+        spinner.dismiss()
         
-        let results : [SearchResult] = self.users.filter({
+        let results : [SearchResult] = users.filter({
             guard let email = $0["email"], email != safeEmail else {
                 return false
             }
@@ -165,17 +165,13 @@ extension NewConversationViewController : UISearchBarDelegate {
     
     func updateUI(){
         if results.isEmpty {
-            self.noResultsLabel.isHidden = false
-            self.tableView.isHidden = true
+            noResultsLabel.isHidden = false
+            tableView.isHidden = true
         } else {
-            self.noResultsLabel.isHidden = true
-            self.tableView.isHidden = false
-            self.tableView.reloadData()
+            noResultsLabel.isHidden = true
+            tableView.isHidden = false
+            tableView.reloadData()
         }
     }
 }
 
-struct SearchResult {
-    let name : String
-    let email : String
-}
