+++ MessagerPractice/Controllers/PhotoViewerViewController.swift
@@ -9,7 +9,7 @@ import UIKit
 
 import SDWebImage
 
-class PhotoViewerViewController: UIViewController {
+final class PhotoViewerViewController: UIViewController {
 
     private let url : URL
     
@@ -33,7 +33,7 @@ class PhotoViewerViewController: UIViewController {
         title = "Photo"
         navigationItem.largeTitleDisplayMode = .never
         view.addSubview(imageView)
-        imageView.sd_setImage(with: self.url, completed: nil)
+        imageView.sd_setImage(with: url, completed: nil)
     }
     
     override func viewDidLayoutSubviews() {
