+++ MessagerPractice/Controllers/ProfileViewController.swift
@@ -11,17 +11,9 @@ import FBSDKLoginKit
 import GoogleSignIn
 import SDWebImage
 
-enum ProfileViewModelType {
-    case info, logout
-}
-struct ProfileViewModel {
-    let viewModelType : ProfileViewModelType
-    let title : String
-    let handler : (() -> ())?
-    
-}
 
-class ProfileViewController: UIViewController {
+
+final class ProfileViewController: UIViewController {
 
     @IBOutlet var tableView : UITableView!
     
@@ -84,7 +76,7 @@ class ProfileViewController: UIViewController {
         
         let path = "images/"+filename
         
-        let headerView = UIView(frame: CGRect(x: 0, y: 0, width: self.view.width, height: 300))
+        let headerView = UIView(frame: CGRect(x: 0, y: 0, width: view.width, height: 300))
         headerView.backgroundColor = .link
         
         let imageView = UIImageView(frame: CGRect(x: (view.width - 150) / 2, y: 75, width: 150, height: 150))
@@ -143,14 +135,14 @@ extension ProfileViewController : UITableViewDelegate, UITableViewDataSource {
 class ProfileTableViewCell : UITableViewCell {
     static let identifier = "ProfileTableViewCell"
     public func setUP(with viewModel : ProfileViewModel){
-        self.textLabel?.text = viewModel.title
+        textLabel?.text = viewModel.title
         switch viewModel.viewModelType {
         case .info:
-            self.textLabel?.textAlignment = .left
-            self.selectionStyle = .none
+            textLabel?.textAlignment = .left
+            selectionStyle = .none
         case .logout:
-            self.textLabel?.textColor = .red
-            self.textLabel?.textAlignment = .center
+            textLabel?.textColor = .red
+            textLabel?.textAlignment = .center
         }
     }
 }
