+++ MessagerPractice/Controllers/SearchResult.swift
@@ -1,0 +1,14 @@
+//
+//  SearchResult.swift
+//  MessagerPractice
+//
+//  Created by 김현미 on 2022/02/27.
+//
+
+import Foundation
+
+struct SearchResult {
+    let name : String
+    let email : String
+}
+
