--- MessagerPractice/Controllers/VideoPlayerViewController.swift
@@ -1,20 +1,0 @@
-//
-//  VideoPlayerViewController.swift
-//  MessagerPractice
-//
-//  Created by jaeseung han on 2022/02/06.
-//
-
-import UIKit
-
-class VideoPlayerViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    
-}
