+++ MessagerPractice/Resources/Extension.swift
@@ -11,26 +11,27 @@ import UIKit
 extension UIView {
     
     public var width : CGFloat {
-        return self.frame.size.width
+        return frame.size.width
     }
     
     public var height : CGFloat {
-        return self.frame.size.height
+        return frame.size.height
     }
     public var top : CGFloat {
-        return self.frame.origin.y
+        return frame.origin.y
     }
     public var bottom : CGFloat {
-        return self.frame.size.height + self.frame.origin.y
+        return frame.size.height + frame.origin.y
     }
     public var left : CGFloat {
-        return self.frame.origin.x
+        return frame.origin.x
     }
     public var right : CGFloat {
-        return self.frame.size.width + self.frame.origin.x
+        return frame.size.width + frame.origin.x
     }
 }
 
 extension Notification.Name {
+    /// Notification when user logs in
     static let didLogInNotification = Notification.Name("didLogInNotification")
 }
