+++ MessagerPractice/Views/ConversationTableViewCell.swift
@@ -52,8 +52,8 @@ class ConversationTableViewCell: UITableViewCell {
     }
     
     public func configure(with model : Conversation){
-        self.userMessageLabel.text = model.latestMessage.text
-        self.userNameLabel.text = model.name
+        userMessageLabel.text = model.latestMessage.text
+        userNameLabel.text = model.name
         
         let path = "images/\(model.otherUserEmail)_profile_picture.png"
         StorageManager.shared.downloadURL(for: path) {[weak self] result in
